Moved two unit tests into it
